Feature: DM evaluation (#13)
Adds a set of functions to compile and run DM code for quick and easy testing.

dm_eval (dmeval, dme) for actually compiling and executing the code.
dm_version (dmversion, dmv) for getting the default BYOND version used.
Heavily inspired by GinjaNinja32's implementation of a similar command with his Bot32. So credit to him on that front.