Overhauls the way logging is handled
Instead of getting a logger once and then passing that logger to modules, the logger is initialized once and each module gets its own logger instance.

The logging is configured in the logging.yml file (which can be changed via the LOG_CFG environment variable)

This changes allow the administrator to configure a different log level + handler per specific module.
For example: the default logging config only logs errors from the discord modules and info logs from the core modules

Also ensures that the channel ids are stored as integer in the channel dictionary (get_channel expects the channel id to be a interger and returns Null otherwise)